Refactor TikTok schemas into submodules
