+++ app/api/tiktok.py
@@ -1,11 +1,7 @@
 from fastapi import APIRouter, status, Request
 from fastapi.responses import JSONResponse
-from app.schemas.tiktok import (
-    TikTokSessionRequest,
-    TikTokSessionResponse,
-    TikTokSearchRequest,
-    TikTokSearchResponse,
-)
+from app.schemas.tiktok.search import TikTokSearchRequest, TikTokSearchResponse
+from app.schemas.tiktok.session import TikTokSessionRequest, TikTokSessionResponse
 from app.services.tiktok.service import TiktokService
