+++ app/services/tiktok/service.py
@@ -8,7 +8,12 @@ from typing import Dict, Any, Optional, Union, List
 from datetime import datetime, timedelta
 from app.services.tiktok.tiktok_executor import TiktokExecutor
 from app.services.tiktok.utils.login_detection import LoginDetector
-from app.schemas.tiktok import TikTokSessionRequest, TikTokSessionResponse, TikTokLoginState, TikTokSessionConfig
+from app.schemas.tiktok.session import (
+    TikTokSessionRequest,
+    TikTokSessionResponse,
+    TikTokLoginState,
+    TikTokSessionConfig,
+)
 from app.core.config import get_settings
 from app.services.tiktok.search_service import TikTokSearchService
